refactor and updated envelope_id on confirm pages
