+++ publishing/management/commands/create_envelope_for_top.py
@@ -1,6 +1,6 @@
 from django.core.management import BaseCommand
 
-from publishing.models import PackagedWorkBasket
+from publishing.models.packaged_workbasket import PackagedWorkBasket
 
 
 class Command(BaseCommand):
