+++ publishing/models/exceptions.py
@@ -1,0 +1,22 @@
+class PackagedWorkBasketDuplication(Exception):
+    pass
+
+
+class PackagedWorkBasketInvalidCheckStatus(Exception):
+    pass
+
+
+class PackagedWorkBasketInvalidQueueOperation(Exception):
+    pass
+
+
+class EnvelopeCurrentlyProccessing(Exception):
+    pass
+
+
+class EnvelopeInvalidQueuePosition(Exception):
+    pass
+
+
+class EnvelopeNoTransactions(Exception):
+    pass
