+++ publishing/tests/test_packaging_queue_views.py
@@ -7,9 +7,9 @@ from bs4 import BeautifulSoup
 from django.urls import reverse
 
 from common.tests import factories
-from publishing.models import OperationalStatus
-from publishing.models import PackagedWorkBasket
-from publishing.models import ProcessingState
+from publishing.models.operational_status import OperationalStatus
+from publishing.models.packaged_workbasket import PackagedWorkBasket
+from publishing.models.state import ProcessingState
 
 pytestmark = pytest.mark.django_db
